fix CLI cutover regressions
Handle optional worker-session inputs, inject Work preparation through Wire, split oversized run files, and reconcile the named lint baselines.
